Refactor MappingAtomicType
