+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/api/types/semtype/MappingAtomicType.java
@@ -19,11 +19,34 @@
 
 package io.ballerina.runtime.api.types.semtype;
 
+import java.util.ArrayList;
+import java.util.Collection;
+
 import static io.ballerina.runtime.api.types.semtype.Builder.CELL_SEMTYPE_INNER_RO;
+import static io.ballerina.runtime.api.types.semtype.Core.cellInner;
+import static io.ballerina.runtime.api.types.semtype.Core.intersectMemberSemTypes;
+import static io.ballerina.runtime.api.types.semtype.Core.isNever;
 
 public record MappingAtomicType(String[] names, SemType[] types, SemType rest) implements AtomicType {
 
     public static final MappingAtomicType MAPPING_ATOMIC_RO = new MappingAtomicType(
             new String[]{}, new SemType[]{}, CELL_SEMTYPE_INNER_RO
     );
+
+    public MappingAtomicType intersectMapping(Env env, MappingAtomicType other) {
+        int expectedSize = Integer.min(types().length, other.types().length);
+        Collection<String> names = new ArrayList<>(expectedSize);
+        Collection<SemType> types = new ArrayList<>(expectedSize);
+        for (FieldPair fieldPair : new FieldPairs(this, other)) {
+            names.add(fieldPair.name());
+            SemType t = intersectMemberSemTypes(env, fieldPair.type1(), fieldPair.type2());
+            if (isNever(cellInner(fieldPair.type1()))) {
+                return null;
+
+            }
+            types.add(t);
+        }
+        SemType rest = intersectMemberSemTypes(env, this.rest(), other.rest());
+        return new MappingAtomicType(names.toArray(String[]::new), types.toArray(SemType[]::new), rest);
+    }
 }
