+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/api/values/BCollection.java
@@ -17,6 +17,8 @@
  */
 package io.ballerina.runtime.api.values;
 
+import io.ballerina.runtime.api.types.semtype.SemType;
+
 /**
  * <p>
  * {@link BCollection} represents a collection in Ballerina.
@@ -32,4 +34,12 @@ public interface BCollection {
      * @return iterator created.
      */
     BIterator<?> getIterator();
+
+    default SemType shapeOf() {
+        return null;
+    }
+
+    default void cacheShape(SemType semType) {
+
+    }
 }
