+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/types/semtype/BMappingSubType.java
@@ -24,24 +24,17 @@ import io.ballerina.runtime.api.types.semtype.Builder;
 import io.ballerina.runtime.api.types.semtype.Conjunction;
 import io.ballerina.runtime.api.types.semtype.Context;
 import io.ballerina.runtime.api.types.semtype.Core;
-import io.ballerina.runtime.api.types.semtype.Env;
+import io.ballerina.runtime.api.types.semtype.FieldPair;
+import io.ballerina.runtime.api.types.semtype.FieldPairs;
 import io.ballerina.runtime.api.types.semtype.MappingAtomicType;
 import io.ballerina.runtime.api.types.semtype.SemType;
 import io.ballerina.runtime.api.types.semtype.SubType;
 
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Iterator;
-import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Objects;
-import java.util.Optional;
 
 import static io.ballerina.runtime.api.types.semtype.Bdd.bddEvery;
 import static io.ballerina.runtime.api.types.semtype.Builder.MAPPING_ATOMIC_INNER;
-import static io.ballerina.runtime.api.types.semtype.Core.cellInner;
-import static io.ballerina.runtime.api.types.semtype.Core.intersectMemberSemTypes;
-import static io.ballerina.runtime.api.types.semtype.Core.isNever;
 
 public class BMappingSubType extends SubType implements DelegatedSubType {
 
@@ -107,7 +100,8 @@ public class BMappingSubType extends SubType implements DelegatedSubType {
                 if (p == null) {
                     break;
                 } else {
-                    MappingAtomicType m = intersectMapping(cx.env, combined, cx.mappingAtomType(p.atom()));
+                    MappingAtomicType m =
+                            combined.intersectMapping(cx.env, cx.mappingAtomType(p.atom()));
                     if (m == null) {
                         return true;
                     } else {
@@ -135,8 +129,7 @@ public class BMappingSubType extends SubType implements DelegatedSubType {
             if (!Core.isEmpty(cx, Core.diff(pos.rest(), neg.rest()))) {
                 return mappingInhabited(cx, pos, negList.next());
             }
-            FieldPairs pairing = new FieldPairs(pos, neg);
-            for (FieldPair fieldPair : pairing) {
+            for (FieldPair fieldPair : new FieldPairs(pos, neg)) {
                 SemType d = Core.diff(fieldPair.type1(), fieldPair.type2());
                 if (!Core.isEmpty(cx, d)) {
                     MappingAtomicType mt;
@@ -144,7 +137,6 @@ public class BMappingSubType extends SubType implements DelegatedSubType {
                         // the posType came from the rest type
                         mt = insertField(pos, fieldPair.name(), d);
                     } else {
-                        // FIXME:
                         if (Core.isSubType(cx, fieldPair.type1(), Builder.cellContaining(cx.env, Builder.undef()))) {
                             continue;
                         }
@@ -188,24 +180,6 @@ public class BMappingSubType extends SubType implements DelegatedSubType {
         return Arrays.copyOf(v, newLength);
     }
 
-    // FIXME: make this an instance method in mappingAtomicType
-    private static MappingAtomicType intersectMapping(Env env, MappingAtomicType m1, MappingAtomicType m2) {
-        int expectedSize = Integer.min(m1.types().length, m2.types().length);
-        List<String> names = new ArrayList<>(expectedSize);
-        List<SemType> types = new ArrayList<>(expectedSize);
-        FieldPairs pairing = new FieldPairs(m1, m2);
-        for (FieldPair fieldPair : pairing) {
-            names.add(fieldPair.name());
-            SemType t = intersectMemberSemTypes(env, fieldPair.type1(), fieldPair.type2());
-            if (isNever(cellInner(fieldPair.type1()))) {
-                return null;
-            }
-            types.add(t);
-        }
-        SemType rest = intersectMemberSemTypes(env, m1.rest(), m2.rest());
-        return new MappingAtomicType(names.toArray(String[]::new), types.toArray(SemType[]::new), rest);
-    }
-
     @Override
     public SubTypeData data() {
         throw new IllegalStateException("unimplemented");
@@ -235,151 +209,4 @@ public class BMappingSubType extends SubType implements DelegatedSubType {
         return Objects.hashCode(inner);
     }
 
-    private static class FieldPairs implements Iterable<FieldPair> {
-
-        MappingAtomicType m1;
-        MappingAtomicType m2;
-        public MappingPairIterator itr;
-
-        public FieldPairs(MappingAtomicType m1, MappingAtomicType m2) {
-            this.m1 = m1;
-            this.m2 = m2;
-        }
-
-        @Override
-        public Iterator<FieldPair> iterator() {
-            itr = new MappingPairIterator(m1, m2);
-            return itr;
-        }
-    }
-
-    private record FieldPair(String name, SemType type1, SemType type2, Integer index1, Integer index2) {
-
-        public static FieldPair create(String name, SemType type1, SemType type2, Integer index1,
-                                       Integer index2) {
-            return new FieldPair(name, type1, type2, index1, index2);
-        }
-    }
-
-    // TODO: refact
-    private static class MappingPairIterator implements Iterator<FieldPair> {
-
-        private final String[] names1;
-        private final String[] names2;
-        private final SemType[] types1;
-        private final SemType[] types2;
-        private final int len1;
-        private final int len2;
-        private int i1 = 0;
-        private int i2 = 0;
-        private final SemType rest1;
-        private final SemType rest2;
-
-        private boolean doneIteration = false;
-        private boolean shouldCalculate = true;
-        private FieldPair cache = null;
-
-        private MappingPairIterator(MappingAtomicType m1, MappingAtomicType m2) {
-            this.names1 = m1.names();
-            this.len1 = this.names1.length;
-            this.types1 = m1.types();
-            this.rest1 = m1.rest();
-            this.names2 = m2.names();
-            this.len2 = this.names2.length;
-            this.types2 = m2.types();
-            this.rest2 = m2.rest();
-        }
-
-        @Override
-        public boolean hasNext() {
-            if (this.doneIteration) {
-                return false;
-            }
-            if (this.shouldCalculate) {
-                FieldPair cache = internalNext();
-                if (cache == null) {
-                    this.doneIteration = true;
-                }
-                this.cache = cache;
-                this.shouldCalculate = false;
-            }
-            return !this.doneIteration;
-        }
-
-        @Override
-        public FieldPair next() {
-            if (this.doneIteration) {
-                throw new NoSuchElementException("Exhausted iterator");
-            }
-
-            if (this.shouldCalculate) {
-                FieldPair cache = internalNext();
-                if (cache == null) {
-                    // this.doneIteration = true;
-                    throw new IllegalStateException();
-                }
-                this.cache = cache;
-            }
-            this.shouldCalculate = true;
-            return this.cache;
-        }
-
-        /*
-         * This method corresponds to `next` method of MappingPairing.
-         */
-        private FieldPair internalNext() {
-            FieldPair p;
-            if (this.i1 >= this.len1) {
-                if (this.i2 >= this.len2) {
-                    return null;
-                }
-                p = FieldPair.create(curName2(), this.rest1, curType2(), null, this.i2);
-                this.i2 += 1;
-            } else if (this.i2 >= this.len2) {
-                p = FieldPair.create(curName1(), curType1(), this.rest2, this.i1, null);
-                this.i1 += 1;
-            } else {
-                String name1 = curName1();
-                String name2 = curName2();
-                if (Common.codePointCompare(name1, name2)) {
-                    p = FieldPair.create(name1, curType1(), this.rest2, this.i1, null);
-                    this.i1 += 1;
-                } else if (Common.codePointCompare(name2, name1)) {
-                    p = FieldPair.create(name2, this.rest1, curType2(), null, this.i2);
-                    this.i2 += 1;
-                } else {
-                    p = FieldPair.create(name1, curType1(), curType2(), this.i1, this.i2);
-                    this.i1 += 1;
-                    this.i2 += 1;
-                }
-            }
-            return p;
-        }
-
-        private SemType curType1() {
-            return this.types1[this.i1];
-        }
-
-        private String curName1() {
-            return this.names1[this.i1];
-        }
-
-        private SemType curType2() {
-            return this.types2[this.i2];
-        }
-
-        private String curName2() {
-            return this.names2[this.i2];
-        }
-
-        public void reset() {
-            this.i1 = 0;
-            this.i2 = 0;
-        }
-
-        public Optional<Integer> index1(String name) {
-            int i1Prev = this.i1 - 1;
-            return i1Prev >= 0 && Objects.equals(this.names1[i1Prev], name) ? Optional.of(i1Prev) : Optional.empty();
-        }
-    }
 }
